PDF-bot-server/requirements.txt
fastapi
uvicorn
python-dotenv
llama-index
shutil
pydantic